fix: allow batching of more than one query inside interactive transactions
Fixes: #25570

This change fixes an issue where running different queries inside an
interactive transaction in the same tick would not be batched together
automatically by the engine.
By accounting for the protocolQuery when batching interactive
transactions, the engine can optimize the queries correctly, as the
queries will be batched by both the transaction ID and the query.